refactor notification component to be replaced by plugins functions
